backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
Flask-Limiter==3.8.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7

# OAuth & Authentication
oauthlib==3.2.2
requests-oauthlib==1.3.0
cryptography==41.0.1

# Production & Security
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.0.0
ecdsa==0.18.0

# Database
psycopg2-binary==2.9.9
redis==5.0.0
alembic==1.13.2
boto3==1.34.162

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.32.0

# Task Queue (for background analytics sync)
celery==5.3.1
kombu==5.3.2

# API Documentation
Flask-RESTX==0.5.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0
